+++ src/Tests/CreateOrderTest.php
@@ -53,20 +53,20 @@ class CreateOrderTest extends TestCase
                     new Address(
                         addressType: 'customer',
                         postalCode: '12345',
-                        street: 'Soborna',
-                        city: 'Poltava',
                         country: new Country(
                             'UA'
-                        )
+                        ),
+                        street: 'Soborna',
+                        city: 'Poltava'
                     ),
                     new Address(
                         addressType: 'billing',
                         postalCode: '1234567',
-                        street: 'Donauweg',
-                        city: 'Amsterdam',
                         country: new Country(
                             'NL'
                         ),
+                        street: 'Donauweg',
+                        city: 'Amsterdam',
                         housenumber: "10"
                     )
                 ),
@@ -119,7 +119,6 @@ class CreateOrderTest extends TestCase
     public function test_sending()
     {
         $response = $this->client->sendOrder($this->order);
-
         self::assertSame($response["status"], 'new');
     }